+++ graphs/views/__init__.py
@@ -1,5 +1,4 @@
 from http import *
-from certificate import *
+from https import *
 from search import *
-# from ip import *
 # from asn import *
